clean up mira-endpoint
