+++ packages/server/src/main/java/software/uncharted/terarium/hmiserver/controller/mira/MiraController.java
@@ -49,11 +49,9 @@ import software.uncharted.terarium.hmiserver.service.data.ProjectService;
 import software.uncharted.terarium.hmiserver.service.tasks.AMRToMMTResponseHandler;
 import software.uncharted.terarium.hmiserver.service.tasks.GenerateModelLatexResponseHandler;
 import software.uncharted.terarium.hmiserver.service.tasks.LatexToAMRResponseHandler;
-import software.uncharted.terarium.hmiserver.service.tasks.LatexToSympyResponseHandler;
 import software.uncharted.terarium.hmiserver.service.tasks.MdlToStockflowResponseHandler;
 import software.uncharted.terarium.hmiserver.service.tasks.SbmlToPetrinetResponseHandler;
 import software.uncharted.terarium.hmiserver.service.tasks.StellaToStockflowResponseHandler;
-import software.uncharted.terarium.hmiserver.service.tasks.SympyToAMRResponseHandler;
 import software.uncharted.terarium.hmiserver.service.tasks.TaskService;
 import software.uncharted.terarium.hmiserver.utils.Messages;
 import software.uncharted.terarium.hmiserver.utils.rebac.Schema;
@@ -252,90 +250,6 @@ public class MiraController {
 		}
 
 		return ResponseEntity.ok().body(latexResponse);
-	}
-
-	@PostMapping("/latex-to-amr")
-	@Secured(Roles.USER)
-	@Operation(summary = "Generate AMR from latex ODE equations")
-	@ApiResponses(
-		value = {
-			@ApiResponse(
-				responseCode = "200",
-				description = "Dispatched successfully",
-				content = @Content(
-					mediaType = "application/json",
-					schema = @io.swagger.v3.oas.annotations.media.Schema(implementation = TaskResponse.class)
-				)
-			),
-			@ApiResponse(responseCode = "500", description = "There was an issue dispatching the request", content = @Content)
-		}
-	)
-	public ResponseEntity<JsonNode> latexToAMR(@RequestBody final String latex) {
-		////////////////////////////////////////////////////////////////////////////////
-		// 1. Convert latex string to python sympy code string
-		//
-		// Note this is a gollm string => string task
-		////////////////////////////////////////////////////////////////////////////////
-		final TaskRequest latexToSympyRequest = new TaskRequest();
-		final TaskResponse latexToSympyResponse;
-		String code = null;
-
-		try {
-			latexToSympyRequest.setType(TaskType.GOLLM);
-			latexToSympyRequest.setInput(latex.getBytes());
-			latexToSympyRequest.setScript(LatexToSympyResponseHandler.NAME);
-			latexToSympyRequest.setUserId(currentUserService.get().getId());
-			latexToSympyResponse = taskService.runTaskSync(latexToSympyRequest);
-
-			final JsonNode node = objectMapper.readValue(latexToSympyResponse.getOutput(), JsonNode.class);
-			code = node.get("response").asText();
-		} catch (final TimeoutException e) {
-			log.warn("Timeout while waiting for task response", e);
-			throw new ResponseStatusException(HttpStatus.SERVICE_UNAVAILABLE, messages.get("task.gollm.timeout"));
-		} catch (final InterruptedException e) {
-			log.warn("Interrupted while waiting for task response", e);
-			throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, messages.get("task.gollm.interrupted"));
-		} catch (final ExecutionException e) {
-			log.error("Error while waiting for task response", e);
-			throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, messages.get("task.gollm.execution-failure"));
-		} catch (final Exception e) {
-			log.error("Unexpected error", e);
-		}
-
-		if (code == null) {
-			throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, messages.get("task.gollm.execution-failure"));
-		}
-
-		////////////////////////////////////////////////////////////////////////////////
-		// 2. Convert python sympy code string to amr
-		//
-		// This returns the AMR json, and intermediate data representations for debugging
-		////////////////////////////////////////////////////////////////////////////////
-		final TaskRequest sympyToAMRRequest = new TaskRequest();
-		final TaskResponse sympyToAMRResponse;
-		final JsonNode response;
-
-		try {
-			sympyToAMRRequest.setType(TaskType.MIRA);
-			sympyToAMRRequest.setInput(code.getBytes());
-			sympyToAMRRequest.setScript(SympyToAMRResponseHandler.NAME);
-			sympyToAMRRequest.setUserId(currentUserService.get().getId());
-			sympyToAMRResponse = taskService.runTaskSync(sympyToAMRRequest);
-			response = objectMapper.readValue(sympyToAMRResponse.getOutput(), JsonNode.class);
-			return ResponseEntity.ok().body(response);
-		} catch (final TimeoutException e) {
-			log.warn("Timeout while waiting for task response", e);
-			throw new ResponseStatusException(HttpStatus.SERVICE_UNAVAILABLE, messages.get("task.mira.timeout"));
-		} catch (final InterruptedException e) {
-			log.warn("Interrupted while waiting for task response", e);
-			throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, messages.get("task.mira.interrupted"));
-		} catch (final ExecutionException e) {
-			log.error("Error while waiting for task response", e);
-			throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, messages.get("task.mira.execution-failure"));
-		} catch (final Exception e) {
-			log.error("Unexpected error", e);
-		}
-		throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, messages.get("generic.io-error.read"));
 	}
 
 	@PostMapping("/convert-and-create-model")
